services/api-gateway/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Authentication and security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Redis client (for rate limiting and caching)
redis==5.0.1
# Redis async support for advanced caching
redis[hiredis]==5.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# OpenTelemetry for distributed tracing
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-httpx==0.41b0
opentelemetry-instrumentation-asyncio==0.41b0
opentelemetry-exporter-jaeger-thrift==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-propagator-b3==1.20.0
opentelemetry-propagator-jaeger==1.20.0

# Additional monitoring dependencies
aiohttp==3.9.1  # For webhook notifications
psutil==5.9.6   # For system metrics

# Performance optimization dependencies
statistics-lib==0.2.2  # For advanced statistics calculations
weakref-utils==1.0.0   # For memory management optimization

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing HTTP client

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment management
python-dotenv==1.0.0